+++ src/torchjd/aggregation/imtl_g.py
@@ -2,6 +2,7 @@ import torch
 from torch import Tensor
 
 from ._gramian_utils import compute_gramian
+from ._non_differentiable import raise_non_differentiable_error
 from .bases import _WeightedAggregator, _Weighting
 
 
@@ -29,6 +30,9 @@ class IMTLG(_WeightedAggregator):
 
     def __init__(self):
         super().__init__(weighting=_IMTLGWeighting())
+
+        # This prevents computing gradients that can be very wrong.
+        self.register_full_backward_pre_hook(raise_non_differentiable_error)
 
 
 class _IMTLGWeighting(_Weighting):